refactor: rename notFound directory and style files to follow kebab-case
