+++ src/components/index.ts
@@ -2,5 +2,5 @@ export { renderCart } from './cart';
 export { renderCheckout } from './checkout';
 export { renderDetails } from './details';
 export { renderHeader } from './header';
-export { renderNotFound } from './notFound';
+export { renderNotFound } from './not-found';
 export { renderProducts } from './products';
