+++ src/styles/checkout/_pop-up.scss
