+++ src/styles/checkout/index.scss
@@ -1,2 +1,2 @@
 @use './forms';
-@use './pop_up';
+@use './pop-up';
