+++ src/styles/index.scss
@@ -9,7 +9,7 @@
 @use 'details';
 @use 'cart';
 @use 'checkout';
-@use 'notFound';
+@use 'not-found';
 
 #root {
   font-size: 1.5rem;
