+++ src/styles/not-found/_not-found.scss
