+++ src/styles/not-found/index.scss
@@ -1,0 +1,1 @@
+@use './not-found';
