--- src/styles/notFound/index.scss
@@ -1,1 +1,0 @@
-@use './not_found';
